xinda-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
sqlalchemy==2.0.36
pillow==11.0.0
python-docx==1.1.2
PyPDF2==3.0.1
PyMuPDF==1.24.14
requests==2.32.3
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.4